% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearner.R
\name{SuperLearner}
\alias{SuperLearner}
\title{SuperLearner}
\description{
SuperLearner

SuperLearner
}
\details{
Ensemble learning class that supports fitting multiple base learners and
a meta-learner using cross-validation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{learner_library}}{Library of base learners.}

\item{\code{X}}{Dataframe or matrix of predictors.}

\item{\code{Y}}{Vector of outcomes.}

\item{\code{learners}}{Base learners used for the SuperLearner.}

\item{\code{fits}}{Fitted base learners.}

\item{\code{meta_learner}}{The meta_learner learned using cross-validation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Superlearner-new}{\code{SuperLearner$new()}}
\item \href{#method-Superlearner-add_learner}{\code{SuperLearner$add_learner()}}
\item \href{#method-Superlearner-fit_learners}{\code{SuperLearner$fit_learners()}}
\item \href{#method-Superlearner-cross_validate}{\code{SuperLearner$cross_validate()}}
\item \href{#method-Superlearner-fit_meta_learner}{\code{SuperLearner$fit_meta_learner()}}
\item \href{#method-Superlearner-predict}{\code{SuperLearner$predict()}}
\item \href{#method-Superlearner-learn}{\code{SuperLearner$learn()}}
\item \href{#method-Superlearner-clone}{\code{SuperLearner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-new"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-new}{}}}
\subsection{Method \code{new()}}{
Initializes the SuperLearner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$new(X, Y, learner_library = c("glmnet", "randomForest", "earth"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Dataframe or matrix of predictors.}

\item{\code{Y}}{Vector of outcomes.}

\item{\code{learner_library}}{library of base learners to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-add_learner"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-add_learner}{}}}
\subsection{Method \code{add_learner()}}{
Manually adds base learners to SuperLearner object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$add_learner(name, fit_fun, predict_fun, x_type = "data.frame")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the base learner.}

\item{\code{fit_fun}}{The fit function of the base learner.}

\item{\code{predict_fun}}{The predict function of the base learner.}

\item{\code{x_type}}{Data type for the predictors, default is data.frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-fit_learners"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-fit_learners}{}}}
\subsection{Method \code{fit_learners()}}{
Fits all base learners using their corresponding fit function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$fit_learners()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-cross_validate"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-cross_validate}{}}}
\subsection{Method \code{cross_validate()}}{
Perform cross-validation using all base learners.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$cross_validate(V = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{V}}{Number of folds for the cross-validation, default is 10.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with n rows and one column per base learner. Each entry
corresponds to the prediction made by a base learner on the held-out sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-fit_meta_learner"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-fit_meta_learner}{}}}
\subsection{Method \code{fit_meta_learner()}}{
Fits the meta learner using glm (minimizes CV-MSE).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$fit_meta_learner(results_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results_matrix}}{The resulting matrix from the cross_validate function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-predict}{}}}
\subsection{Method \code{predict()}}{
Computes prediction on predictors using the meta-learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A dataframe or matrix of predictors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-learn"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-learn}{}}}
\subsection{Method \code{learn()}}{
Computes the full stack of functions to learn the meta-learner and prints
the glm output for the meta-learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$learn(V = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{V}}{Number of folds for cross-validation, default is 10}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Superlearner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Superlearner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SuperLearner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
