% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{TMLE}
\alias{TMLE}
\title{TMLE}
\description{
TMLE

TMLE
}
\details{
TMLE class that computes ATE and confidence interval of the ATE assuming
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{W_A}}{Dataframe or matrix of predictors (includes covariates W and treatments A).}

\item{\code{Y}}{Vector of outcomes.}

\item{\code{outcome_superLearner}}{SuperLearner used for expected outcome predictions.}

\item{\code{propensity_superLearner}}{SuperLearner used for propensity score predictions.}

\item{\code{treatment_name}}{Name of the treatment column in W_A.}

\item{\code{full_tibble}}{Tibble containing all relevant measures.}

\item{\code{fit}}{glm fit of the estimating equation.}

\item{\code{eps}}{Coefficient of clever covariate of the fit.}

\item{\code{tmle_ate}}{ATE value estimation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TMLE-new}{\code{TMLE$new()}}
\item \href{#method-TMLE-compute_ATE}{\code{TMLE$compute_ATE()}}
\item \href{#method-TMLE-clone}{\code{TMLE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-new"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-new}{}}}
\subsection{Method \code{new()}}{
Initilizes a TMLE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TMLE$new(
  W_A,
  Y,
  outcome_superLearner,
  propensity_superLearner,
  treatment_name = "A"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{W_A}}{Dataframe or matrix of predictors (includes covariates W and treatments A).}

\item{\code{Y}}{Vector of outcomes.}

\item{\code{outcome_superLearner}}{SuperLearner used for expected outcome predictions.}

\item{\code{propensity_superLearner}}{SuperLearner used for propensity score predictions.}

\item{\code{treatment_name}}{Name of the treatment column in W_A.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-compute_ATE"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-compute_ATE}{}}}
\subsection{Method \code{compute_ATE()}}{
Computes the ATE and confidence interval of the ATE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TMLE$compute_ATE()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list containing estimated_ATE as the estimation of the ATE,
ci_low as the lower bound of the confidence interval and ci_high as the
upper bound of the confidence interval.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
